feat: Switched to development configuration in Docker and updated Rails model specs
The switch from a production to a development environment in the Docker configuration promotes a more flexible development experience. We have also added a 'credentials/development.key' to Gitignore to prevent the inclusion of confidential information in the code repository. These changes, coupled with enhancing the Dockerfile by adding 'libpq-dev' and 'postgresql-client', allow for better support of PostgreSQL.

The Rails' database.yml file also saw a migration from a flexible environment variable-based configuration to a fixed one to streamline the local development setup process. The addition of 'dotenv' gem ensures easy management of environment variables in the development setting.

Model specs for Evaluated, InstrumentApplication, Instrument, and Psychologist saw a significant expansion. We have added more test cases for better coverage and validation of relationships between these models. These additions solidify our commitment to Test-Driven Development (TDD) while improving confidence in code accuracy and functionality.

Furthermore, a new 'ApplicationRecord' file is added to establish a common base class for all models which is an encouraged best-practice by Rails.

This commit also includes a database schema that provides a detailed representation of the database structure, further strengthening foundational knowledge of the application.