feat: Add psychologist authentication and evaluatee email notification
Authentication feature has been implemented for psychologists to enhance security.
Now, a valid psychologist must be logged in to access the application, otherwise the user will be
gently urged to sign in first. This required a modification in the application controller.

Meanwhile, an interaction flow for evaluated users has also been introduced; allowing them to
start instrument evaluation upon entering their details and create a record for themselves.
This includes sending an email notification to the evaluated user whenever a new psychological
instrument becomes available for them. New pages and templates have been added to facilitate this,
as seen in the diff for views.

Several tests have been added to ensure the smooth functioning of the new evaluation flow and
psychologist authentication thus maintaining the robustness of the application.

Lastly, the `rails-controller-testing` gem was introduced to support new testing requirements.